+++ src/main/java/org/yt/jr/projects/crypto/BruteForce.java
@@ -1,6 +1,6 @@
 package org.yt.jr.projects.crypto;
 
-import org.yt.jr.projects.alphabet.Alphabet;
+import org.yt.jr.projects.lang.Alphabet;
 import java.util.HashMap;
 
 public class BruteForce {
